fix(auth): extract escapeHtml to html.ts, stop logging raw OAuth error
Addresses two review follow-ups on #123:

1. Extract escapeHtml into src/auth/html.ts. Importing src/auth/server.ts
   is NOT side-effect-free — module load runs resolveOAuthScopes(), which
   reads GOOGLE_DRIVE_MCP_SCOPES and throws on an unknown alias, and pulls
   in express/google-auth-library/open. The escapeHtml test now imports a
   hermetic pure-function module instead.

2. Log error.message only in the /oauth2callback catch, not the raw error.
   google-auth-library token-exchange failures are GaxiosErrors whose
   config/response carry the request body (authorization code +
   client_secret); console.error(error) would write those to logs. Also
   relabel "token exchange failed" -> "OAuth callback failed" since the
   try block also covers token persistence.
