+++ src/auth/html.ts
@@ -1,0 +1,10 @@
+// Escape user-influenced strings before embedding them in HTML responses.
+// Covers both element content and double-quoted attribute contexts.
+export function escapeHtml(s: string): string {
+  return s
+    .replace(/&/g, '&amp;')
+    .replace(/</g, '&lt;')
+    .replace(/>/g, '&gt;')
+    .replace(/"/g, '&quot;')
+    .replace(/'/g, '&#39;');
+}
